Part 1
A DNS query contains 2 parts: a header and a question.
The header has:
a. a query ID
b. some flags
c. 4 counts

The question has:
a. name - the domain name
b. DNS record type
c. a class (always the same)

Part 1 focuses on encoding the name in question to a binary sequence, which is the format of a DNS query.
With that encoded name, it built the DNS query and used an UDP socket to send the query to Google's DNS resolver to test it.